Fix 64-bit cleanliness bug
Casting a native pointer to (uint32) is technically the wrong thing
to do when calculating alignment fixups. You should really use
uintptr_t, otherwise the code is not 64-bit clean. Technically
this doesn't here because we intend this code to run on a Cortex-M7
core which is only 32-bit, but I want this bug documented somewhere
in case anybody else ever wonders why Nofrendo segfaults when you
try to compile and run it on their 64-bit machine (like it did for
me with FreeBSD/amd64).

+++ software/firmware/nofrendo-2.0pre1/src/bitmap.c
@@ -25,6 +25,7 @@
 
 #include <stdio.h>
 #include <string.h>
+#include <stdint.h>
 #include <noftypes.h>
 #include <bitmap.h>
 
@@ -61,7 +62,8 @@ static bitmap_t *_make_bitmap(uint8 *data_addr, bool hw, int width,
    if (false == bitmap->hardware)
    {
       bitmap->pitch = (bitmap->pitch + 3) & ~3;
-      bitmap->line[0] = (uint8 *) (((uint32) bitmap->data + overdraw + 3) & ~3);
+      bitmap->line[0] = (uint8 *) (((uintptr_t) bitmap->data +
+        overdraw + 3) & ~3);
    }
    else
    {
